refactor output block
